--- a/codeowners.go
+++ b/codeowners.go
@@ -225,21 +225,14 @@ func (c *Codeowners) Owners(path string) []string {
 
 // precompile all regular expressions
 var (
-	reCommentIgnore = regexp.MustCompile(`^(\\#|\\!)`)
-	rePrependSlash  = regexp.MustCompile(`([^\/+])/.*\*\.`)
-	reEscapeDot     = regexp.MustCompile(`\.`)
-	reDoubleStar1   = regexp.MustCompile(`/\*\*/`)
-	reDoubleStar2   = regexp.MustCompile(`\*\*/`)
-	reDoubleStar3   = regexp.MustCompile(`/\*\*`)
-	reEscapeStar1   = regexp.MustCompile(`\\\*`)
-	reEscapeStar2   = regexp.MustCompile(`\*`)
+	rePrependSlash = regexp.MustCompile(`([^\/+])/.*\*\.`)
 )
 
 // based on github.com/sabhiram/go-gitignore
 // but modified so that 'dir/*' only matches files in 'dir/'
 func getPattern(line string) (*regexp.Regexp, error) {
 	// when # or ! is escaped with a \
-	if reCommentIgnore.MatchString(line) {
+	if strings.HasPrefix(line, `\#`) || strings.HasPrefix(line, `\!`) {
 		line = line[1:]
 	}
 
@@ -249,7 +242,7 @@ func getPattern(line string) (*regexp.Regexp, error) {
 	}
 
 	// Handle escaping the "." char
-	line = reEscapeDot.ReplaceAllString(line, `\.`)
+	line = strings.ReplaceAll(line, ".", `\.`)
 
 	magicStar := "#$~"
 
@@ -257,13 +250,13 @@ func getPattern(line string) (*regexp.Regexp, error) {
 	if strings.HasPrefix(line, "/**/") {
 		line = line[1:]
 	}
-	line = reDoubleStar1.ReplaceAllString(line, `(/|/.+/)`)
-	line = reDoubleStar2.ReplaceAllString(line, `(|.`+magicStar+`/)`)
-	line = reDoubleStar3.ReplaceAllString(line, `(|/.`+magicStar+`)`)
+	line = strings.ReplaceAll(line, `/**/`, `(/|/.+/)`)
+	line = strings.ReplaceAll(line, `**/`, `(|.`+magicStar+`/)`)
+	line = strings.ReplaceAll(line, `/**`, `(|/.`+magicStar+`)`)
 
 	// Handle escaping the "*" char
-	line = reEscapeStar1.ReplaceAllString(line, `\`+magicStar)
-	line = reEscapeStar2.ReplaceAllString(line, `([^/]*)`)
+	line = strings.ReplaceAll(line, `\*`, `\`+magicStar)
+	line = strings.ReplaceAll(line, `*`, `([^/]*)`)
 
 	// Handle escaping the "?" char
 	line = strings.ReplaceAll(line, "?", `\?`)