This PR reduces the allocations in `getPattern` and it also improves the overall speed of the function.

First note that I had to fix the existing benchmark, as the reader was being read only once in the first iteration, and afterwards it was always empty, thus `parseCodeowners` was parsing an empty file.

These were the metrics before the change:

    $ go test -benchmem -bench=.
    goos: darwin
    goarch: arm64
    pkg: github.com/hairyhenderson/go-codeowners
    BenchmarkParseCodeowners-10      2677503               465.3 ns/op          4096 B/op          1 allocs/op
    PASS
    ok      github.com/hairyhenderson/go-codeowners 2.255s

After the change it's not showing more realistic data:

    $ go test -benchmem -bench=.
    goos: darwin
    goarch: arm64
    pkg: github.com/hairyhenderson/go-codeowners
    BenchmarkParseCodeowners-10        31054             38910 ns/op           61424 B/op        641 allocs/op
    PASS
    ok      github.com/hairyhenderson/go-codeowners 3.529s

Then we move out of `getPattern` all regexp compilations, reducing the allocations ~62% and the performance in ~55%. Finally uses `strings.ReplaceAll` instead of the regexp versions when possible. These are the end results:

```
$ benchstat main.log stringsreplace.log
goos: darwin
goarch: arm64
pkg: github.com/hairyhenderson/go-codeowners
                   │  main.log   │         stringsreplace.log          │
                   │   sec/op    │   sec/op     vs base                │
ParseCodeowners-10   38.72µ ± 2%   15.21µ ± 7%  -60.73% (p=0.000 n=10)

                   │   main.log   │          stringsreplace.log          │
                   │     B/op     │     B/op      vs base                │
ParseCodeowners-10   59.95Ki ± 0%   25.98Ki ± 0%  -56.66% (p=0.000 n=10)

                   │  main.log  │         stringsreplace.log         │
                   │ allocs/op  │ allocs/op   vs base                │
ParseCodeowners-10   641.0 ± 0%   194.0 ± 0%  -69.73% (p=0.000 n=10)
```