@inkel 2024-09-04
I honestly don't know what the CI error is about 🤔