@hairyhenderson 2024-09-25
@inkel you'll need to amend the commit message(s) to match conventional commit patterns - maybe something like "fix(perf): Reduce allocations"